requirements.txt
# Core
python-dotenv==1.0.1
pydantic>=2.0.0,<3.0.0
pyyaml>=6.0,<7.0

# LLM
litellm>=1.55.0,<2.0.0
openai>=1.0.0,<2.0.0
anthropic>=0.40.0,<1.0.0

# Vector DB & Embeddings
chromadb>=0.5.0,<1.0.0
sentence-transformers>=3.0.0,<4.0.0

# HTTP & Async
httpx>=0.25.0,<1.0.0
aiohttp>=3.9.0,<4.0.0

# Text processing
jieba>=0.42.1

# State Machine
python-statemachine>=3.0.0,<4.0.0

# CLI
rich>=13.0.0,<14.0.0
click>=8.1.0,<9.0.0

# Dev
pytest>=7.0.0,<9.0.0
pytest-asyncio>=0.21.0,<1.0.0
